+++ lib/omq/routing/client.rb
@@ -9,19 +9,19 @@ module OMQ
     class Client
       include RoundRobin
 
+
+      # @return [Async::LimitedQueue]
+      #
+      attr_reader :recv_queue
+
+
       # @param engine [Engine]
       #
       def initialize(engine)
         @engine     = engine
         @recv_queue = Routing.build_queue(engine.options.recv_hwm, :block)
-        @tasks      = []
         init_round_robin(engine)
       end
-
-
-      # @return [Async::LimitedQueue]
-      #
-      attr_reader :recv_queue
 
 
       # Dequeues the next received message. Blocks until one is available.
@@ -46,8 +46,7 @@ module OMQ
       #
       def connection_added(connection)
         @connections << connection
-        task = @engine.start_recv_pump(connection, @recv_queue)
-        @tasks << task if task
+        @engine.start_recv_pump(connection, @recv_queue)
         add_round_robin_send_connection(connection)
       end
 
@@ -66,12 +65,6 @@ module OMQ
         enqueue_round_robin(parts)
       end
 
-
-      # Stops all background tasks.
-      def stop
-        @tasks.each(&:stop)
-        @tasks.clear
-      end
     end
   end
 end