+++ lib/omq/routing/dealer.rb
@@ -20,7 +20,6 @@ module OMQ
       def initialize(engine)
         @engine     = engine
         @recv_queue = Routing.build_queue(engine.options.recv_hwm, :block)
-        @tasks      = []
         init_round_robin(engine)
       end
 
@@ -46,8 +45,7 @@ module OMQ
       # @param connection [Connection]
       #
       def connection_added(connection)
-        task = @engine.start_recv_pump(connection, @recv_queue)
-        @tasks << task if task
+        @engine.start_recv_pump(connection, @recv_queue)
         add_round_robin_send_connection(connection)
       end
 
@@ -64,16 +62,6 @@ module OMQ
       #
       def enqueue(parts)
         enqueue_round_robin(parts)
-      end
-
-
-      # Stops all background tasks.
-      #
-      # @return [void]
-      #
-      def stop
-        @tasks.each(&:stop)
-        @tasks.clear
       end
 
     end