+++ lib/omq/routing/pub.rb
@@ -15,7 +15,6 @@ module OMQ
       #
       def initialize(engine)
         @engine = engine
-        @tasks  = []
         init_fan_out(engine)
       end
 
@@ -60,16 +59,6 @@ module OMQ
       #
       def enqueue(parts)
         fan_out_enqueue(parts)
-      end
-
-
-      # Stops all background tasks.
-      #
-      # @return [void]
-      #
-      def stop
-        @tasks.each(&:stop)
-        @tasks.clear
       end
 
     end