+++ lib/omq/routing/push.rb
@@ -12,7 +12,6 @@ module OMQ
       #
       def initialize(engine)
         @engine = engine
-        @tasks  = []
         init_round_robin(engine)
       end
 
@@ -57,14 +56,6 @@ module OMQ
       end
 
 
-      # Stops all background tasks (send pumps, reapers).
-      #
-      def stop
-        @tasks.each(&:stop)
-        @tasks.clear
-      end
-
-
       private
 
 
@@ -74,7 +65,7 @@ module OMQ
       #
       def start_reaper(conn)
         return if conn.is_a?(Transport::Inproc::DirectPipe)
-        @tasks << @engine.spawn_conn_pump_task(conn, annotation: "reaper") do
+        @engine.spawn_conn_pump_task(conn, annotation: "reaper") do
           conn.receive_message # blocks until peer disconnects; then exits
         end
       end